fix: unify windows-transparent into windows
