--- cradle/windows-transparent/.gitignore
@@ -1,3 +1,0 @@
-src/userlib
-src/userlib.rs
-Cargo.toml
