--- cradle/windows-transparent/Cargo.template.toml
@@ -1,23 +1,0 @@
-[package]
-name = "peridot-cradle-windows"
-version = "0.1.0"
-authors = ["S.Percentage <Syn.Tri.Naga@gmail.com>"]
-edition = "2018"
-build = "build.rs"
-
-[features]
-UseExternalAssetPath = []
-
-[dependencies]
-bedrock = { git = "https://github.com/Pctg-x8/bedrock", branch = "peridot", features = ["Implements", "VK_KHR_external_memory_win32", "VK_KHR_external_semaphore_win32"] }
-peridot = { path = "../.." }
-log = "0.4"
-env_logger = "0.5"
-comdrive = { git = "https://github.com/Pctg-x8/comdrive" }
-metrics = { git = "https://github.com/Pctg-x8/metrics" }
-widestring = "0.4"
-#%KERNEL_CRATE_NAME% = { path = "%KERNEL_CRATE_PATH%" }
-
-[dependencies.winapi]
-version = "0.3"
-features = ["winuser", "libloaderapi", "shellscalingapi", "handleapi", "synchapi", "winbase"]
