--- cradle/windows-transparent/build.rs
@@ -1,3 +1,0 @@
-fn main() {
-    println!("cargo:rustc-link-search=static={}/Lib", env!("VK_SDK_PATH"));
-}
